+++ lib/modules/data/tracker/queries/tracked_entity_instance.query.dart
@@ -1,8 +1,8 @@
 import 'package:dhis2_flutter_sdk/core/annotations/index.dart';
 import 'package:dhis2_flutter_sdk/core/utilities/repository.dart';
-import 'package:dhis2_flutter_sdk/modules/data/tracked_entity_instance/entities/enrollment.entity.dart';
-import 'package:dhis2_flutter_sdk/modules/data/tracked_entity_instance/entities/tracked-entity.entity.dart';
-import 'package:dhis2_flutter_sdk/modules/data/tracked_entity_instance/entities/tracked_entity_attribute_value.entity.dart';
+import 'package:dhis2_flutter_sdk/modules/data/tracker/entities/enrollment.entity.dart';
+import 'package:dhis2_flutter_sdk/modules/data/tracker/entities/tracked-entity.entity.dart';
+import 'package:dhis2_flutter_sdk/modules/data/tracker/entities/tracked_entity_attribute_value.entity.dart';
 import 'package:dhis2_flutter_sdk/shared/queries/base.query.dart';
 import 'package:reflectable/mirrors.dart';
 import 'package:sqflite/sqflite.dart';
@@ -28,7 +28,8 @@ class TrackedEntityInstanceQuery extends BaseQuery<TrackedEntityInstance> {
                   as ClassMirror),
           referencedEntityColumns: Entity.getEntityColumns(
               AnnotationReflectable.reflectType(TrackedEntityAttributeValue)
-                  as ClassMirror));
+                  as ClassMirror,
+              false));
       this.relations.add(relation);
     }
 
@@ -49,7 +50,8 @@ class TrackedEntityInstanceQuery extends BaseQuery<TrackedEntityInstance> {
           referencedEntity: Entity.getEntityDefinition(
               AnnotationReflectable.reflectType(Enrollment) as ClassMirror),
           referencedEntityColumns: Entity.getEntityColumns(
-              AnnotationReflectable.reflectType(Enrollment) as ClassMirror));
+              AnnotationReflectable.reflectType(Enrollment) as ClassMirror,
+              false));
       this.relations.add(relation);
     }