+++ test/repository_test.reflectable.dart
@@ -783,7 +783,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             22,
             23,
             const <int>[3],
-            const <int>[33],
+            const <int>[33, 34],
             const prefix0.AnnotationReflection(),
             const []),
         r.ImplicitGetterMirrorImpl(
@@ -802,11 +802,11 @@ final _data = <r.Reflectable, r.ReflectorData>{
             2,
             2,
             const <int>[],
-            const <int>[34, 35, 36, 37],
+            const <int>[35, 36, 37, 38],
             const prefix0.AnnotationReflection(),
             const []),
         r.MethodMirrorImpl(r'getType', 131090, 3, -1, 9, 9, const <int>[],
-            const <int>[38], const prefix0.AnnotationReflection(), const []),
+            const <int>[39], const prefix0.AnnotationReflection(), const []),
         r.MethodMirrorImpl(
             r'getColumnType',
             131090,
@@ -815,7 +815,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             24,
             24,
             const <int>[],
-            const <int>[39],
+            const <int>[40],
             const prefix0.AnnotationReflection(),
             const []),
         r.MethodMirrorImpl(
@@ -826,7 +826,7 @@ final _data = <r.Reflectable, r.ReflectorData>{
             3,
             3,
             const <int>[],
-            const <int>[40, 41],
+            const <int>[41, 42, 43],
             const prefix0.AnnotationReflection(),
             const []),
         r.ImplicitGetterMirrorImpl(
@@ -874,11 +874,11 @@ final _data = <r.Reflectable, r.ReflectorData>{
             3,
             3,
             const <int>[],
-            const <int>[42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53],
+            const <int>[44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55],
             const prefix0.AnnotationReflection(),
             const []),
         r.MethodMirrorImpl(r'', 128, 4, -1, 4, 4, const <int>[],
-            const <int>[54], const prefix0.AnnotationReflection(), const []),
+            const <int>[56], const prefix0.AnnotationReflection(), const []),
         r.MethodMirrorImpl(r'', 128, 5, -1, 5, 5, const <int>[], const <int>[],
             const prefix0.AnnotationReflection(), const []),
         r.ImplicitGetterMirrorImpl(
@@ -895,13 +895,13 @@ final _data = <r.Reflectable, r.ReflectorData>{
             6,
             6,
             const <int>[],
-            const <int>[55, 56, 57],
+            const <int>[57, 58, 59],
             const prefix0.AnnotationReflection(),
             const []),
         r.ImplicitGetterMirrorImpl(
             const prefix0.AnnotationReflection(), 33, 96),
         r.MethodMirrorImpl(r'', 128, 7, -1, 7, 7, const <int>[],
-            const <int>[58], const prefix0.AnnotationReflection(), const [])
+            const <int>[60], const prefix0.AnnotationReflection(), const [])
       ],
       <m.ParameterMirror>[
         r.ParameterMirrorImpl(
@@ -1313,6 +1313,18 @@ final _data = <r.Reflectable, r.ReflectorData>{
             null,
             null),
         r.ParameterMirrorImpl(
+            r'ignoreRelationColumns',
+            32774,
+            66,
+            const prefix0.AnnotationReflection(),
+            -1,
+            13,
+            13,
+            const <int>[],
+            const [],
+            null,
+            null),
+        r.ParameterMirrorImpl(
             r'tableName',
             41990,
             71,
@@ -1404,6 +1416,18 @@ final _data = <r.Reflectable, r.ReflectorData>{
             -1,
             9,
             9,
+            const <int>[],
+            const [],
+            null,
+            null),
+        r.ParameterMirrorImpl(
+            r'ignoreRelationColumns',
+            32774,
+            74,
+            const prefix0.AnnotationReflection(),
+            -1,
+            13,
+            13,
             const <int>[],
             const [],
             null,