+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-## [1.0.0-alpha.1] - 06/01/2022
+## [1.0.0-alpha.4] - 29/01/2022
 
-- Support for online authenitication
+- Support to save tracked entity instances together with their attributes and enrollment details
+
+## [1.0.0-alpha.3] - 28/01/2022
+
+- Support to fetch user organisation unit
 
 ## [1.0.0-alpha.2] - 17/01/2022
 
 - Support for downloading metadata
+
+## [1.0.0-alpha.1] - 06/01/2022
+
+- Support for online authenitication